requirements.txt
# Core
python-dotenv>=1.0.0
openai>=1.12.0

# GUI
customtkinter>=5.2.0
tkinterdnd2>=0.3.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Image Processing
pillow>=10.0.0

# Utilities
lxml>=4.9.0
html5lib>=1.1
markdown>=3.5.0
